requirements.txt
accelerate
beautifulsoup4
faster-whisper
imageio
imageio-ffmpeg
matplotlib
numpy
openai
openai-whisper
opencv-python
pandas
playwright
python-dotenv
seaborn
statsmodels
tqdm
torch
transformers
torchvision
